feat: 훅으로 리팩터링
