+++ src/pages/Home.tsx
@@ -59,6 +59,7 @@ import { useAuth } from '@/hooks/useAuth';
 import useWebsocket, { connectToWebSockets, disconnectWebSockets } from '@/hooks/useWebsocket';
 import { Lesson } from '@/types/learning';
 import { useGlobalWebSocketStatus } from '@/contexts/GlobalWebSocketContext';
+import { useChapterHandler } from '@/hooks/useChapterHandler';
 
 
 const { Search: AntdSearch } = Input;
@@ -174,7 +175,7 @@ const Dashboard: React.FC = () => {
   // 온보딩 및 손 선호도 모달 상태
   const [isHandPreferenceModalOpen, setIsHandPreferenceModalOpen] = useState(false);
   const [shouldShowOnboarding, setShouldShowOnboarding] = useState(false);
-  const [connectingChapter, setConnectingChapter] = useState<string | null>(null);
+
   // 전체 진도율 원형 그래프 변수 선언 (JSX 바깥에서)
   const percent = progressOverview?.overall_progress || 0;
   const radius = 56;
@@ -322,87 +323,7 @@ const Dashboard: React.FC = () => {
     navigate(`/learn/word/${encodeURIComponent(selectedItem)}`);
   };
 
-  const handleStartLearn = async (chapterId: string, lessonIds: string[]) => {
-    const modeNum = 1;
-    const path = `/learn/chapter/${chapterId}/guide/${modeNum}`;
-    try {
-      setConnectingChapter(chapterId);
-
-      // 1. 챕터 프로그레스 초기화 API 호출 (user_chapter_progress, user_lesson_progress 생성)
-      await API.post(`/progress/chapters/${chapterId}`);
-
-      // 2. WebSocket 연결 시도
-      try {
-        const response = await API.get<{ success: boolean; data: { ws_urls: string[], lesson_mapper: { [key: string]: string } } }>(`/ml/deploy/${chapterId}`);
-        if (response.data.success && response.data.data.ws_urls) {
-          console.log('[Chapters]response.data.data.lesson_mapper', response.data.data.lesson_mapper);
-          await connectToWebSockets(response.data.data.ws_urls);
-          showStatus(); // 전역 상태 표시 활성화
-
-          // 학습 진도 이벤트 기록
-          await API.post('/progress/lessons/events', { lesson_ids: lessonIds, mode: 'study' });
-
-          // lesson_mapper를 URL state로 전달
-          navigate(path, {
-            state: {
-              lesson_mapper: response.data.data.lesson_mapper
-            }
-          });
-          return; // 성공적으로 처리되었으므로 함수 종료
-        }
-      } catch (wsError) {
-        console.warn('WebSocket 연결 실패:', wsError);
-        // WebSocket 연결 실패해도 페이지 이동은 계속 진행
-      }
-
-      setConnectingChapter(null);
-      navigate(path);
-    } catch (err) {
-      console.error('학습 시작 실패:', err);
-      setConnectingChapter(null);
-      navigate(path); // 실패해도 이동
-    }
-  };
-
-  const handleStartQuiz = async (chapterId: string, lessonIds: string[]) => {
-    const modeNum = 2;
-    const path = `/learn/chapter/${chapterId}/guide/${modeNum}`;
-    try {
-      setConnectingChapter(chapterId);
-
-      // WebSocket 연결 시도
-      try {
-        const response = await API.get<{ success: boolean; data: { ws_urls: string[], lesson_mapper: { [key: string]: string } } }>(`/ml/deploy/${chapterId}`);
-        if (response.data.success && response.data.data.ws_urls) {
-          console.log('[Chapters]response.data.data.lesson_mapper', response.data.data.lesson_mapper);
-          await connectToWebSockets(response.data.data.ws_urls);
-          showStatus(); // 전역 상태 표시 활성화
-
-          // 학습 진도 이벤트 기록
-          await API.post('/progress/lessons/events', { lesson_ids: lessonIds, mode: 'review' });
-
-          // lesson_mapper를 URL state로 전달
-          navigate(path, {
-            state: {
-              lesson_mapper: response.data.data.lesson_mapper
-            }
-          });
-          return; // 성공적으로 처리되었으므로 함수 종료
-        }
-      } catch (wsError) {
-        console.warn('WebSocket 연결 실패:', wsError);
-        // WebSocket 연결 실패해도 페이지 이동은 계속 진행
-      }
-
-      setConnectingChapter(null);
-      navigate(path);
-    } catch (err) {
-      console.error('학습 시작 실패:', err);
-      setConnectingChapter(null);
-      navigate(path); // 실패해도 이동
-    }
-  };
-
+  const { connectingChapter, handleStartLearn, handleStartQuiz } = useChapterHandler();
 
   const handleCardClick = (cardType: string) => {
     switch (cardType) {
